+++ luadatapacktypes/src/source.d.ts
@@ -1,4 +1,4 @@
-/// <reference path='vec3d.d.ts'/>
+/// <reference path='vec.d.ts'/>
 /// <reference path='entity.d.ts'/>
 
 /**